requirements.txt
fastapi
uvicorn
pydantic
numpy
pyyaml
boto3
mysql-connector-python
requests